bugfix on gcc:
the gcc compiler does not deduce the correct function in the    ->decltype(do_end3(p, unfold_end(p.t))) statement.

if the statement is removed it works fine but auto returntypes are only avalible in c++ 14 therefore it needs to bin in there for CLING for the time beeing. this means cling cannot use async at the moment